feat(compose): add rememberHDREnvironment, rename IBLEnvironment to Environment
rememberHDREnvironment builds an Environment from a raw equirectangular
.hdr (no cmgen bake), prefiltering the skybox and reflections on the GPU
at load via IBLPrefilter. Rename the IBLEnvironment type to Environment
so it reads consistently across the KTX and HDR source loaders.
